Added new function gauss_kernel_3d() which produces a 3D Gaussian kernel for smoothing in 3D.
Modified function tophat_kernel_3d() so that it produces a 3D tophat kernel
either cubical or spherical. The option shape controls the latter, default is
cubical. The previous version of this function produced a 3D arrays of 1s so
it did not work well.

All tabs replaced by spaces (emacs)

 Changes to be committed:
	modified:   smoothing.py